ensured placeholders are created/discarded in the DOM when re-rendering
previously obsolete placeholders were still present in the DOM while new
placeholders were not being added because tiles no longer or not yet in
the matrix were being ignored

so we now remove all tiles from the DOM, then add them back in as we
traverse the matrix - this is a bit crude though and likely causes
memory leaks (cf. inline comment)

this also ensures that the transform - i.e. the eventual position - is
always relative to the top-left corner (previously the dragged item was
not being animated because its eventual position was set immediately)